backend/ballotBackend/requirements.txt
django 
python-dotenv
djangorestframework
django-storages
django-cors-headers
deepface
pinecone-client